Adds new arg vars to test mod
