+++ src/core/tests.pri
@@ -6,3 +6,4 @@ SOURCES += \
 	$$PWD/tcpstreamtest.cpp \
 	$$PWD/unixstreamtest.cpp \
 	$$PWD/eventlooptest.cpp
+
